+++ src/chunker/mod.rs
@@ -1,6 +1,5 @@
-mod Fq12MUL;
 pub mod assigner;
 pub mod common;
 pub mod elements;
-pub mod fq12mul;
+pub mod fq12_mul;
 pub mod segment;
